requirements.txt
# Core ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# Processamento
pyspark>=3.5.0

# Orquestracao
prefect>=2.14.0

# Tracking
mlflow>=2.9.0

# Validacao
great-expectations>=0.18.0

# Serving
fastapi>=0.104.0
uvicorn>=0.24.0

# Monitoramento
evidently>=0.4.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utils
python-dotenv>=1.0.0
joblib>=1.3.0
httpx>=0.25.0
requests>=2.31.0
pyarrow>=14.0.0

# Testes
pytest>=7.4.0
pytest-cov>=4.1.0

# Lint
ruff>=0.1.0
